refactor: moved exported function
